+++ app/Repositories/CityRepositoryInterface.php
@@ -2,7 +2,7 @@
 
 namespace App\Repositories;
 
-use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Support\Collection;
 
 interface CityRepositoryInterface
 {