+++ app/Repositories/ProductRepositoryInterface.php
@@ -6,7 +6,7 @@ use Illuminate\Database\Eloquent\Collection;
 
 interface ProductRepositoryInterface
 {
-    public function getProducts(int $userId, ?array $filterProducts = null): array;
+    public function getProductsBySellers(array $sellerIds, ?array $filterProducts = null): array;
     public function getProductId(string $productType): Collection;
     public function setProduct(array $data): int;
 }