+++ resources/views/partials/city-filter.blade.php
@@ -3,8 +3,8 @@
 <select class="selectpicker show-tick" data-style="btn-primary" data-live-search="true" id="city" name="city">
     <option value="" selected="selected">Todas</option>
     <option data-divider="true"></option>
-    @foreach($cities as $city)
-        <option value='{{ $city->id }}'>{{ $city->name }}</option>
+    @foreach($cities as $id => $name)
+        <option value='{{ $id }}'>{{ $name }}</option>
     @endforeach
 </select>
 <hr />